rag-pipeline/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
langchain==0.0.208
pinecone-client==2.2.1
